refactor: user/nameCheck -> auth/nameCheck
사유 : auth 안에 있어야 authGuard에 걸리지 않음
